Optimize algorithm for winner calculation
Prior to this change, for each move the program would loop through all
possible (8) winning combinations.

This change opimizes the algorithm in such a way that no more than 4
combinations is required (3, on an average) to be checked, based on the
particular move.

However, this optimization comes at a cost: memory. I believe it now
requires about three times the memory. But hey, at least it's about
twice as fast!